pyproject.toml
[tool.poetry]
name = "voice-conversation-ai"
version = "0.1.0"
description = ""
authors = ["kill_in_sun <noneck.0311@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
soundfile = "^0.12.1"
speechrecognition = "^3.10.3"
mlx = "^0.12.0"
tiktoken = "^0.6.0"
huggingface = "^0.0.1"
scipy = "^1.13.0"
pyaudio = "^0.2.14"
huggingface-hub = "^0.22.2"
numba = "^0.59.1"
ollama = "^0.1.9"
openai = "^1.23.6"
python-dotenv = "^1.0.1"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
